+++ sdk/storagediscovery/arm-storagediscovery/samples/v1-beta/typescript/src/storageDiscoveryWorkspacesCreateOrUpdateSample.ts
@@ -26,7 +26,7 @@ async function createOrUpdateAStorageDiscoveryWorkspace(): Promise<void> {
         capabilities: {
           azureBlobStorage: {
             capacityDetails: { status: "Enabled" },
-            prefixDefinitions: [
+            prefixConfigurations: [
               {
                 storageAccountName: "firstsa",
                 containerName: "samplecontainer",